Move sanitize_name to util
